# public/swagger.yaml
openapi: 3.0.3
info:
  title: FBZ Barbearia API
  description: |
    API completa para gerenciamento de barbearia incluindo agendamentos, barbeiros, clientes, produtos, planos e webhooks.

    ## Autenticação
    A maioria dos endpoints requer autenticação via Bearer Token no header:
    ```
    Authorization: Bearer SEU_TOKEN_AQUI
    ```

    ## Formato de Datas
    - Datas são no formato brasileiro: DD/MM/YYYY ou YYYY-MM-DD
    - Horários são no formato: HH:MM (24h)

    ## Webhooks
    O sistema suporta dois tipos de webhooks:
    1. **Webhook Global**: Configurado na tabela `configuracoes`, dispara para todos os eventos
    2. **Webhooks de Barbeiros**: Webhooks personalizados por profissional

  version: 3.0.0
  contact:
    name: FBZ Barbearia
    url: https://fbzbarbearia.vercel.app
servers:
  - url: https://fbzbarbearia.vercel.app/api
    description: Produção
  - url: http://localhost:3000/api
    description: Desenvolvimento
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token de autenticação da API
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Mensagem de erro
        errors:
          type: array
          items:
            type: string
...
paths:
...
  /clientes/deletar:
    delete:
      tags:
        - Clientes
      summary: Deletar cliente (incluindo leads frios)
      description: |
        Deleta permanentemente um cliente do sistema.

        **⚠️ ATENÇÃO:** Esta operação é irreversível!

        **Validações:**
        - Se o cliente possuir agendamentos futuros (status: agendado/confirmado), a deleção será BLOQUEADA
        - Se o cliente possuir apenas histórico ou agendamentos concluidos/cancelados, será deletado normalmente
        - Ideal para uso por IA/chatbots para remover leads frios automaticamente

        **Quando usar:**
        - Leads que nunca agendaram após X dias
        - Clientes inativos que desejam ser removidos
        - Contas duplicadas que precisam ser mergeadas (deletar a duplicada)

        **Casos de uso por IA:**
        ```bash
        # Deletar lead frio por ID
        curl -X DELETE 'https://fbzbarbearia.vercel.app/api/clientes/deletar' \
          -H 'Authorization: Bearer SEU_TOKEN' \
          -H 'Content-Type: application/json' \
          -d '{
            "cliente_id": "uuid-cliente-aqui"
          }'

        # Deletar lead frio por telefone
        curl -X DELETE 'https://fbzbarbearia.vercel.app/api/clientes/deletar' \
          -H 'Authorization: Bearer SEU_TOKEN' \
          -H 'Content-Type: application/json' \
          -d '{
            "telefone": "11999999999"
          }'
        ```
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - cliente_id
              properties:
                cliente_id:
                  type: string
                  format: uuid
                  description: UUID do cliente a ser deletado (OBRIGATÓRIO se não usar telefone)
                  example: "550e8400-e29b-41d4-a716-446655440000"
                telefone:
                  type: string
                  description: Telefone do cliente (OBRIGATÓRIO se não usar cliente_id)
                  example: "11999999999"
            examples:
              por_id:
                summary: Deletar por ID (Recomendado)
                value:
                  cliente_id: "550e8400-e29b-41d4-a716-446655440000"
              por_telefone:
                summary: Deletar por telefone
                value:
                  telefone: "11999999999"
      responses:
        '200':
          description: Cliente deletado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Cliente deletado com sucesso!"
                  data:
                    type: object
                    properties:
                      cliente_id:
                        type: string
                        format: uuid
                      nome:
                        type: string
                        example: "João Silva"
                      telefone:
                        type: string
                        example: "11999999999"
                      deletado_em:
                        type: string
                        format: date-time
        '400':
          description: Não é possível deletar o cliente
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Não é possível deletar este cliente"
                  motivo:
                    type: string
                    example: "Cliente possui agendamentos futuros agendados"
                  agendamentos_futuros:
                    type: integer
                    example: 2
                    description: Quantidade de agendamentos futuros que impedem a deleção
                  detalhes:
                    type: string
                    example: "Cancele ou conclua todos os agendamentos antes de deletar o cliente"
        '401':
          description: Token não fornecido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Token inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Cliente não encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Cliente não encontrado"
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...